Revert "There were a number of things that needed fixing to run"
This reverts commit 18d61cb580d349036b351fe3bbf6a1a8400a0d0e.

The build and dev run will work with this change on Node 16.x.x.

Testing still will not run. It gets an error trying to find
babel-preset-env, which as far as I can tell should not be a thing.

I tried installing the babel-present-env package just to see how it
affected things.  It got rid of that error and replaced it with
new ones.